refactored report to have common template context. Adapted reportEmail to use inherited context.
